+++ samples/Catalog/examples/EventListeners.js
@@ -14,16 +14,8 @@ export class EventListeners extends BaseExampleAutoHidingHeaderComponent {
     this.pdfRef = React.createRef();
 
     hideToolbar(navigation);
-
-    navigation.addListener('beforeRemove', e => {
-      this.pdfRef?.current?.destroyView();
-    });
   }
 
-  componentWillUnmount() {
-    const { navigation } = this.props;
-    navigation.removeListener('beforeRemove');
-  }
   render() {
     return (
       <View style={styles.flex}>