+++ samples/Catalog/examples/GeneratePDF.js
@@ -3,24 +3,13 @@ import { BaseExampleAutoHidingHeaderComponent } from '../helpers/BaseExampleAuto
 import { processColor, View } from 'react-native';
 import PSPDFKitView from 'react-native-pspdfkit';
 import { pspdfkitColor } from '../configuration/Constants';
-import { hideToolbar } from '../helpers/NavigationHelper';
 
 export class GeneratePDF extends BaseExampleAutoHidingHeaderComponent {
   pdfRef = null;
 
   constructor(props) {
     super(props);
-    const { navigation } = this.props;
     this.pdfRef = React.createRef();
-
-    navigation.addListener('beforeRemove', e => {
-      this.pdfRef?.current?.destroyView();
-    });
-  }
-
-  componentWillUnmount() {
-    const { navigation } = this.props;
-    navigation.removeListener('beforeRemove');
   }
 
   render() {