+++ samples/Catalog/examples/ProgrammaticFormFilling.js
@@ -8,18 +8,9 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
 
   constructor(props) {
     super(props);
-    const { navigation } = this.props;
     this.pdfRef = React.createRef();
-
-    navigation.addListener('beforeRemove', e => {
-      this.pdfRef?.current?.destroyView();
-    });
   }
 
-  componentWillUnmount() {
-    const { navigation } = this.props;
-    navigation.removeListener('beforeRemove');
-  }
   render() {
     return (
       <View style={styles.flex}>