+++ samples/NativeCatalog/examples/GeneratePDF.js
@@ -1,6 +1,6 @@
 import { BaseExampleAutoHidingHeaderComponent } from '../helpers/BaseExampleAutoHidingHeaderComponent';
-import { CustomPdfView, formDocumentPath } from '../configuration/Constants';
-import { Platform, View } from 'react-native';
+import { CustomPdfView } from '../configuration/Constants';
+import { View } from 'react-native';
 import React from 'react';
 
 export class GeneratePDF extends BaseExampleAutoHidingHeaderComponent {
@@ -10,26 +10,6 @@ export class GeneratePDF extends BaseExampleAutoHidingHeaderComponent {
     this.state = {
       shouldReturn: false,
     };
-
-    const { navigation } = this.props;
-
-    navigation.addListener('beforeRemove', e => {
-      if (Platform.OS !== 'android') {
-        return;
-      }
-      const { shouldReturn } = this.state;
-      if (!shouldReturn) {
-        this.setState({ shouldReturn: true });
-        e.preventDefault();
-      }
-      setTimeout(() => {
-        this.goBack();
-      }, 50);
-    });
-  }
-
-  goBack(navigation) {
-    this.props.navigation.goBack();
   }
 
   render() {